const OpcUa_StringA is incorrect
Several places in the code (most prominently in opcua_string.c, but in
various other places too) used "const OpcUa_StringA" with the apparent
intent to be a pointer to constant memory.  However that's not what that
does, that gives a constant pointer to mutable memory.  In contrast,
"const OpcUa_String*" is fine because OpcUa_String is a struct.

Discussion:
https://stackoverflow.com/questions/11619236/how-to-typedef-a-constant-variable
https://cboard.cprogramming.com/c-programming/167494-issue-pointer-const-char-using-typedef.html
https://stackoverflow.com/questions/8504411/typedef-pointer-const-weirdness

Solution is to define a type that specifically means a pointer to
constant memory.  This is why, for example, Windows defines PCSTR in
addition to PSTR.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>